Divide loaders/actions of team pages
