+++ app/features/team/loaders/t.$customUrl.roster.server.ts
@@ -1,0 +1,29 @@
+import type { LoaderFunctionArgs } from "@remix-run/node";
+import { redirect } from "@remix-run/node";
+import { requireUserId } from "~/features/auth/core/user.server";
+import { isAdmin } from "~/permissions";
+import { notFoundIfFalsy } from "~/utils/remix.server";
+import { teamPage } from "~/utils/urls";
+import * as TeamRepository from "../TeamRepository.server";
+import { teamParamsSchema } from "../team-schemas.server";
+import { isTeamManager } from "../team-utils";
+import "../team.css";
+
+export const loader = async ({ request, params }: LoaderFunctionArgs) => {
+	const user = await requireUserId(request);
+	const { customUrl } = teamParamsSchema.parse(params);
+
+	const team = notFoundIfFalsy(
+		await TeamRepository.findByCustomUrl(customUrl, {
+			includeInviteCode: true,
+		}),
+	);
+
+	if (!isTeamManager({ team, user }) && !isAdmin(user)) {
+		throw redirect(teamPage(customUrl));
+	}
+
+	return {
+		team,
+	};
+};
